tests/wpt/web-platform-tests/css/css-values/calc-invalid-parsing.html

<!doctype html>
<title>Invalid calc() expressions</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support/parsing-testcommon.js"></script>
<link rel="help" href="https://drafts.csswg.org/css-values-4/#calc-syntax">
<script>
test_invalid_value('transform', 'rotate(calc((0.25turn error)))');
test_invalid_value('width', 'calc(7px * up)');
test_invalid_value('width', 'round(nearest, 1px, 1px, 1px)');
test_invalid_value('width', 'round(nearest, 1px)');
</script>
